Reorganise classes now that structure is somewhat rigid
