+++ tests/test_causal_problem/test_cpcomponent.py
@@ -5,11 +5,8 @@ import numpy.typing as npt
 import pytest
 from numpyro.handlers import condition, do
 
-from causalprog.causal_problem.causal_estimand import (
-    HandlerToApply,
-    Model,
-    _CPComponent,
-)
+from causalprog.causal_problem.components import _CPComponent
+from causalprog.causal_problem.handlers import HandlerToApply, Model
 from causalprog.graph import Graph
